Fase 2 ok

+++ src/main/java/br/com/tokiomarine/seguradora/avaliacao/service/EstudanteServiceImpl.java
@@ -2,6 +2,7 @@ package br.com.tokiomarine.seguradora.avaliacao.service;
 
 import java.util.List;
 import java.util.NoSuchElementException;
+import java.util.Optional;
 
 import javax.validation.Valid;
 
@@ -14,44 +15,49 @@ import org.springframework.stereotype.Service;
 @Service
 public class EstudanteServiceImpl implements EstudandeService {
 
-	@Autowired
-	private EstudanteRepository repository;
+    @Autowired
+    private EstudanteRepository repository;
 
-	@Override
-	public void cadastrarEstudante(@Valid Estudante estudante) {
-		repository.saveAndFlush(estudante);
-	}
+    @Override
+    public void cadastrarEstudante(@Valid Estudante estudante) {
+        repository.saveAndFlush(estudante);
+    }
 
-	@Override
-	public void atualizarEstudante(@Valid Estudante estudante) {
-		Estudante tmp = repository.getById(estudante.getId()).get();
-		tmp.setNome(estudante.getNome());
-		tmp.setEmail(estudante.getEmail());
-		tmp.setTelefone(estudante.getTelefone());
-		repository.saveAndFlush(tmp);
-	}
+    @Override
+    public void atualizarEstudante(@Valid Estudante estudante) {
+        Estudante tmp = repository.getById(estudante.getId()).get();
+        tmp.setNome(estudante.getNome());
+        tmp.setEmail(estudante.getEmail());
+        tmp.setTelefone(estudante.getTelefone());
+        tmp.setMatricula(estudante.getMatricula());
+        tmp.setCurso(estudante.getCurso());
+        repository.saveAndFlush(tmp);
+    }
 
-	@Override
-	public List<Estudante> buscarEstudantes() {
-		return repository.findAll();
-	}
+    @Override
+    public List<Estudante> buscarEstudantes() {
+        return repository.findAll();
+    }
 
-	@Override
-	public Estudante buscarEstudante(Integer id) {
-		return repository.findById(id).get();
-	}
+    @Override
+    public Estudante buscaPorMatricula(String matricula) {
+        return repository.getByMatricula(matricula).orElse(null);
+    }
 
-	@Override
-	public void deletar(Integer id) {
-		try {
-			Estudante tmp = repository.findById(id).get();
-			repository.delete(tmp);
-		}catch (NoSuchElementException nex){
-			System.out.println("Elemento do id: "+id+" não encontrado");
-		}catch (Exception ex){
-			ex.printStackTrace();
-		}
-
+    @Override
+    public Estudante buscarEstudante(Integer id) {
+		Optional<Estudante> optional = repository.findById(id);
+		return optional.orElse(null);
 
 	}
+
+    @Override
+    public void deletar(Integer id) {
+        Optional<Estudante> optional = repository.findById(id);
+        if (optional.isPresent()) {
+            Estudante tmp = optional.get();
+            repository.delete(tmp);
+        }
+
+    }
 }
